refactor(gitignore): added ffmpeg library to .gitignore)
